[FIX/#354] 학교, 학과 검색 페이지 Int로 수정

+++ data/src/main/java/com/example/data/repository/OnboardingRepositoryImpl.kt
@@ -32,8 +32,7 @@ class OnboardingRepositoryImpl @Inject constructor(
 
     override suspend fun getSchoolList(
         keyword: String,
-        page: Long,
-
+        page: Int,
     ): Result<SchoolList?> {
         return runCatching {
             onboardingDataSource.getSchoolNameData(
@@ -43,7 +42,10 @@ class OnboardingRepositoryImpl @Inject constructor(
         }
     }
 
-    override suspend fun getHighSchoolList(keyword: String, page: Long): Result<HighSchoolList?> {
+    override suspend fun getHighSchoolList(
+        keyword: String,
+        page: Int,
+    ): Result<HighSchoolList?> {
         return runCatching {
             onboardingDataSource.getHighSchoolNameData(
                 keyword,
